bug feed punti esperienza

+++ ModelloDati/Marte5BeautifulVinoModelloDati/src/com/marte5/modello2/Feed.java
@@ -70,7 +70,7 @@ public class Feed {
 	 * @return the puntiEsperienza
 	 */
 	@DynamoDBAttribute(attributeName="puntiEsperienza")
-	public long getPuntiEsperienza() {
+	public int getPuntiEsperienza() {
 		return puntiEsperienza;
 	}
 	/**
